build: change version to not include edition
This change removes the edition suffix (-ce/-ee) from the version string in both the frontend and backend. It was
originally included because Sentry requires different version names for the two built artifacts, so I added a separate
SENTRY_VERSION environment variable for the frontend that does include the suffix. The edition is now shown separately
in the backend and frontend which makes more sense anyways, IMO.

Signed-off-by: Jakob Steiner <jakob.steiner@glasskube.eu>